tests: move f number tests to test_atom
